Split starknet-contract compiler into multiple files
